added logic for download files
